split InitSetup into useInitSetup hook + view, extract pure persistence
